Refactor BasicFrame Move DesignPanel into separate file
